Move GameModes to shared/ for cross-package use
